Declare arguments for memstr.h operations as pointers instead of uintptr_t.

+++ kernel/arch/amd64/src/pm.c
@@ -155,7 +155,7 @@ void idt_setoffset(idescriptor_t *d, uintptr_t offset)
 
 void tss_initialize(tss_t *t)
 {
-	memsetb((uintptr_t) t, sizeof(tss_t), 0);
+	memsetb(t, sizeof(tss_t), 0);
 }
 
 /*
@@ -239,7 +239,7 @@ void arch_reboot(void)
 	preemption_disable();
 	ipl_t ipl = interrupts_disable();
 	
-	memsetb((uintptr_t) idt, sizeof(idt), 0);
+	memsetb(idt, sizeof(idt), 0);
 	idtr_load(&idtr);
 	
 	interrupts_restore(ipl);
